requirements_feature_pipeline.txt
confluent-kafka
hopsworks
packaging
pandas
pyarrow
python-dotenv
requests
holidays
numpy
hsfs
scikit-learn==1.5.2
